requirements.txt
langchain>=0.3
langchain-openai>=0.2
langchain-core>=0.3
langchain-community>=0.3
faiss-cpu>=1.7
openai>=1.0
streamlit>=1.28
tiktoken>=0.5
pypdf>=3.0
python-docx>=1.0
numpy>=1.24
python-dotenv